3D
This was by far my simplest project, as I was working in 3D space. It's very difficult, and although I could've done a lot more with it, there were many issues with my code which made it nigh impossible to progress, and it's much harder to visualise than 2D space, which is what I'm used to working with. I plan to work more with 3D space in the future as it's very fun to work with, but nevertheless difficult.