3D Space
I had a lot of trouble with this project. It ended up being very simple, since I couldn't grasp 3D space very well. I encountered some problems and couldn't really see what the issue was since the ball was bouncing up and down somewhere far away from me which was difficult to see whether it was getting farther or whether it was getting lower or higher or whatever the case.